pasado a double y con medicion de benchamark

+++ PuntosMasCercanos/src/main/kotlin/Punto.kt
@@ -1,4 +1,4 @@
 data class Punto(
-    val x:Float,
-    val y:Float
+    val x:Double,
+    val y:Double
 )
